requirements.txt
maskpass
Office365-REST-Python-Client
